+++ net/server.go
@@ -2,11 +2,9 @@ package net
 
 import (
 	"blockchain_go/blockchain"
-	"blockchain_go/miner"
 	"blockchain_go/transaction"
 	"bytes"
 	"encoding/gob"
-	"encoding/hex"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -22,8 +20,6 @@ var nodeAddress string
 var miningAddress string
 var CenterNode string = "localhost:3000"
 var knownNodes = []string{CenterNode}
-var blocksInTransit = [][]byte{}
-var mempool = make(map[string]transaction.Transaction)
 
 type addr struct {
 	AddrList []string
@@ -171,224 +167,6 @@ func sendVersion(addr string, bc *blockchain.Blockchain) {
 	request := append(commandToBytes("version"), payload...)
 
 	sendData(addr, request)
-}
-
-func handleAddr(request []byte) {
-	var buff bytes.Buffer
-	var payload addr
-
-	buff.Write(request[commandLength:])
-	dec := gob.NewDecoder(&buff)
-	err := dec.Decode(&payload)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	knownNodes = append(knownNodes, payload.AddrList...)
-	fmt.Printf("There are %d known nodes now!\n", len(knownNodes))
-	requestBlocks()
-}
-
-func handleBlock(request []byte, bc *blockchain.Blockchain) {
-	var buff bytes.Buffer
-	var payload block
-
-	buff.Write(request[commandLength:])
-	dec := gob.NewDecoder(&buff)
-	err := dec.Decode(&payload)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	blockData := payload.Block
-	block := blockchain.DeserializeBlock(blockData)
-
-	fmt.Println("Recevied a new block!")
-	bc.AddBlock(block)
-
-	fmt.Printf("Added block %x\n", block.Hash)
-
-	if len(blocksInTransit) > 0 {
-		blockHash := blocksInTransit[0]
-		sendGetData(payload.AddrFrom, "block", blockHash)
-
-		blocksInTransit = blocksInTransit[1:]
-	} else {
-		UTXOSet := blockchain.UTXOSet{bc}
-		UTXOSet.Reindex()
-	}
-}
-
-func handleInv(request []byte, bc *blockchain.Blockchain) {
-	var buff bytes.Buffer
-	var payload inv
-
-	buff.Write(request[commandLength:])
-	dec := gob.NewDecoder(&buff)
-	err := dec.Decode(&payload)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	fmt.Printf("Recevied inventory with %d %s\n", len(payload.Items), payload.Type)
-
-	if payload.Type == "block" {
-		blocksInTransit = payload.Items
-
-		blockHash := payload.Items[0]
-		sendGetData(payload.AddrFrom, "block", blockHash)
-
-		newInTransit := [][]byte{}
-		for _, b := range blocksInTransit {
-			if bytes.Compare(b, blockHash) != 0 {
-				newInTransit = append(newInTransit, b)
-			}
-		}
-		blocksInTransit = newInTransit
-	}
-
-	if payload.Type == "tx" {
-		txID := payload.Items[0]
-
-		if mempool[hex.EncodeToString(txID)].ID == nil {
-			sendGetData(payload.AddrFrom, "tx", txID)
-		}
-	}
-}
-
-func handleGetBlocks(request []byte, bc *blockchain.Blockchain) {
-	var buff bytes.Buffer
-	var payload getblocks
-
-	buff.Write(request[commandLength:])
-	dec := gob.NewDecoder(&buff)
-	err := dec.Decode(&payload)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	blocks := bc.GetBlockHashes()
-	sendInv(payload.AddrFrom, "block", blocks)
-}
-
-func handleGetData(request []byte, bc *blockchain.Blockchain) {
-	var buff bytes.Buffer
-	var payload getdata
-
-	buff.Write(request[commandLength:])
-	dec := gob.NewDecoder(&buff)
-	err := dec.Decode(&payload)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	if payload.Type == "block" {
-		block, err := bc.GetBlock([]byte(payload.ID))
-		if err != nil {
-			return
-		}
-
-		sendBlock(payload.AddrFrom, &block)
-	}
-
-	if payload.Type == "tx" {
-		txID := hex.EncodeToString(payload.ID)
-		tx := mempool[txID]
-
-		SendTx(payload.AddrFrom, &tx)
-		// delete(mempool, txID)
-	}
-}
-
-func handleTx(request []byte, bc *blockchain.Blockchain) {
-	var buff bytes.Buffer
-	var payload tx
-
-	buff.Write(request[commandLength:])
-	dec := gob.NewDecoder(&buff)
-	err := dec.Decode(&payload)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	txData := payload.Transaction
-	tx := transaction.DeserializeTransaction(txData)
-	mempool[hex.EncodeToString(tx.ID)] = tx
-
-	if nodeAddress == knownNodes[0] {
-		for _, node := range knownNodes {
-			if node != nodeAddress && node != payload.AddFrom {
-				sendInv(node, "tx", [][]byte{tx.ID})
-			}
-		}
-	} else {
-		if len(mempool) >= 2 && len(miningAddress) > 0 {
-		MineTransactions:
-			var txs []*transaction.Transaction
-
-			for id := range mempool {
-				tx := mempool[id]
-				if bc.VerifyTransaction(&tx) {
-					txs = append(txs, &tx)
-				}
-			}
-
-			if len(txs) == 0 {
-				fmt.Println("All transactions are invalid! Waiting for new ones...")
-				return
-			}
-
-			cbTx := transaction.NewCoinbaseTX(miningAddress, "")
-			txs = append(txs, cbTx)
-
-			newBlock := miner.MineBlock(bc, txs)
-			UTXOSet := blockchain.UTXOSet{bc}
-			UTXOSet.Reindex()
-
-			fmt.Println("New block is mined!")
-
-			for _, tx := range txs {
-				txID := hex.EncodeToString(tx.ID)
-				delete(mempool, txID)
-			}
-
-			for _, node := range knownNodes {
-				if node != nodeAddress {
-					sendInv(node, "block", [][]byte{newBlock.Hash})
-				}
-			}
-
-			if len(mempool) > 0 {
-				goto MineTransactions
-			}
-		}
-	}
-}
-
-func handleVersion(request []byte, bc *blockchain.Blockchain) {
-	var buff bytes.Buffer
-	var payload version
-
-	buff.Write(request[commandLength:])
-	dec := gob.NewDecoder(&buff)
-	err := dec.Decode(&payload)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	myBestHeight := bc.GetBestHeight()
-	foreignerBestHeight := payload.BestHeight
-
-	if myBestHeight < foreignerBestHeight {
-		sendGetBlocks(payload.AddrFrom)
-	} else if myBestHeight > foreignerBestHeight {
-		sendVersion(payload.AddrFrom, bc)
-	}
-
-	// sendAddr(payload.AddrFrom)
-	if !nodeIsKnown(payload.AddrFrom) {
-		knownNodes = append(knownNodes, payload.AddrFrom)
-	}
 }
 
 func handleConnection(conn net.Conn, bc *blockchain.Blockchain) {
